+++ xcube/util/zarrinsert.py
@@ -34,93 +34,78 @@ def check_append_or_insert(time_range, output_path):
         t_center = (t1 + t2) / 2
     else:
         t_center = t1
-    ds = xr.open_zarr(output_path)
-    if np.greater(t_center, get_time_in_days_since_1970(ds.time.values[-1])):
-        # append modus
-        return True
-    else:
-        return False
+    with xr.open_zarr(output_path) as ds:
+        return np.greater(t_center, get_time_in_days_since_1970(ds.time.values[-1]))
 
 
-def merge_single_zarr_into_destination_zarr(src_path, dst_path):
+def insert_input_file_into_output_path(input_path, output_path):
     """Merging the data for the new time stamp into the existing and remaining zarr directory."""
-    merge = _check_if_unique(src_path, dst_path)
-    if merge:
-        ds_single = xr.open_zarr(src_path)
-        ds = xr.open_zarr(dst_path)
-        src_idx = 0
-        new_idx = np.amin(np.where(ds.time[:] > (ds_single.time.values[src_idx])))
-        ds.close()
-        ds_single.close()
+    insert = _check_if_insert_into_output_path(input_path, output_path)
+    if insert:
+        with xr.open_zarr(input_path) as input_ds, xr.open_zarr(output_path) as output_ds:
+            input_time_idx = 0
+            new_time_idx = np.amin(np.where(output_ds.time[:] > (input_ds.time.values[input_time_idx])))
         # Preparing the source directory with the single time stamp to be ready for merging
         # --> files of variables, excluding "lat" and "lon" need to be renamed
-        _rename_file(src_path, src_idx, new_idx)
-        # Preparing the destination directory to be ready for single time stam to be merged
-        # --> files of variables, excluding "lat" and "lon" need to be renamed
-        # The renaming needs to happen in reversed order and starting at the index of nearest above value:
-        for i in reversed(range(new_idx, ds.time.shape[0])):
-            _rename_file(dst_path, i, (i + 1))
-        # Final step: copy the single time stamp files into the destination zarr and adjusting .zarray to the change.
-        _copy_into_target(src_path, dst_path, new_idx)
-        return True
+            _rename_file(input_path, input_time_idx, new_time_idx)
+            # Preparing the destination directory to be ready for single time stam to be merged
+            # --> files of variables, excluding "lat" and "lon" need to be renamed
+            # The renaming needs to happen in reversed order and starting at the index of nearest above value:
+            for i in reversed(range(new_time_idx, output_ds.time.shape[0])):
+                _rename_file(output_path, i, (i + 1))
+            # Final step: copy the single time stamp files into the destination zarr and adjusting .zarray to the change.
+            _copy_into_output_path(input_path, output_path, new_time_idx)
+            return True
     else:
         return False
 
 
-def _check_if_unique(src_path, dst_path):
+def _check_if_insert_into_output_path(input_path, output_path):
     """Check if to be added time stamp is unique """
-    ds = xr.open_zarr(dst_path)
-    ds_single = xr.open_zarr(src_path)
-    mask = np.equal(ds.time.values, ds_single.time.values)
-    ds.close()
-    ds_single.close()
-    if not mask.any():
-        return True
-    else:
-        return False
+    with xr.open_zarr(input_path) as input_ds, xr.open_zarr(output_path) as output_ds:
+        mask = np.equal(output_ds.time.values, input_ds.time.values)
+        return not mask.any()
 
 
-def _rename_file(path_to_ds, old_index, new_time_i):
+def _rename_file(path, old_time_idx, new_time_idx):
     """Renaming files within the directories according to new time index."""
-    ds = xr.open_zarr(path_to_ds)
-    variables = ds.variables
-    ds.close()
+    with xr.open_zarr(path) as ds:
+        variables = ds.variables
     for v in variables:
         if (v != 'lat') and (v != 'lon') and (v != 'lat_bnds') and (v != 'lon_bnds'):
-            path = os.path.join(path_to_ds, v)
-            for root, dirs, files in os.walk(path):
+            v_path = os.path.join(path, v)
+            for root, dirs, files in os.walk(v_path):
                 for filename in files:
                     parts1 = filename.split('.', 1)[0]
-                    if parts1 == (str(old_index)) and (v != "time"):
+                    if parts1 == (str(old_time_idx)) and (v != "time"):
                         parts2 = filename.split('.', 1)[1]
-                        new_name = (str(new_time_i) + '.{}').format(parts2)
-                        os.rename(os.path.join(path, filename), os.path.join(path, new_name))
-                    elif parts1 == (str(old_index)) and (v == "time"):
-                        os.rename(os.path.join(path, filename), os.path.join(path, str(new_time_i)))
+                        new_name = (str(new_time_idx) + '.{}').format(parts2)
+                        os.rename(os.path.join(v_path, filename), os.path.join(v_path, new_name))
+                    elif parts1 == (str(old_time_idx)) and (v == "time"):
+                        os.rename(os.path.join(v_path, filename), os.path.join(v_path, str(new_time_idx)))
 
 
-def _copy_into_target(src_path, dst_path, src_index):
+def _copy_into_output_path(input_path, output_path, input_time_idx):
     """Copy the files with the new time stamp into the existing zarr directory."""
-    ds = xr.open_zarr(src_path)
-    variables = ds.variables
-    ds.close()
-    for v in variables:
-        if (v != 'lat') and (v != 'lon') and (v != 'lat_bnds') and (v != 'lon_bnds'):
-            path = os.path.join(src_path, v)
-            for root, dirs, files in os.walk(path):
+    with xr.open_zarr(input_path) as input_ds:
+        variables = input_ds.variables
+    for variable in variables:
+        if (variable != 'lat') and (variable != 'lon') and (variable != 'lat_bnds') and (variable != 'lon_bnds'):
+            v_path = os.path.join(input_path, variable)
+            for root, dirs, files in os.walk(v_path):
                 for filename in files:
                     parts1 = filename.split('.', 1)[0]
-                    if parts1 == str(src_index):
-                        shutil.copyfile((os.path.join(src_path, v, filename)), (os.path.join(dst_path, v, filename)))
-            _adjust_zarray(dst_path, v)
+                    if parts1 == str(input_time_idx):
+                        shutil.copyfile((os.path.join(input_path, variable, filename)), (os.path.join(output_path, variable, filename)))
+            _adjust_zarray(output_path, variable)
 
 
-def _adjust_zarray(dst_path, variable):
+def _adjust_zarray(output_path, variable):
     """Changing the shape for time in the .zarray file."""
-    with open((os.path.join(dst_path, variable, '.zarray')), 'r') as jsonFile:
+    with open((os.path.join(output_path, variable, '.zarray')), 'r') as jsonFile:
         data = json.load(jsonFile)
     t_shape = data["shape"]
     data["shape"][0] = t_shape[0] + 1
 
-    with open((os.path.join(dst_path, variable, '.zarray')), 'w') as jsonFile:
+    with open((os.path.join(output_path, variable, '.zarray')), 'w') as jsonFile:
         json.dump(data, jsonFile, indent=4)